Bugfix/1067 (#1447)
* Bugfix #1067

* Bugfix and removed obsolete migrations #1067

* Bugfixes #1067

* Bugfixes #1067

+++ webclient/pas/signals.py
@@ -35,21 +35,28 @@ def save_metadata_samostatny_nalez(sender, instance: SamostatnyNalez, **kwargs):
 
 
 @receiver(pre_delete, sender=SamostatnyNalez)
-def dokument_delete_soubor_vazby(sender, instance: SamostatnyNalez, **kwargs):
-    try:
-        instance.soubory.delete()
-    except ObjectDoesNotExist:
-        logger.debug("pas.signals.dokument_delete_soubor_vazby.not_exist", extra={"ident_cely": instance.ident_cely})
-
-@receiver(pre_delete, sender=SamostatnyNalez)
-def samostatny_nalez_okres_delete_repository_container(sender, instance: SamostatnyNalez, **kwargs):
+def dokument_delete_container_soubor_vazby(sender, instance: SamostatnyNalez, **kwargs):
+    logger.debug("pas.signals.dokument_delete_container_soubor_vazby.start",
+                 extra={"ident_cely": instance.ident_cely})
     instance.record_deletion()
+    if instance.soubory and instance.soubory.pk:
+        instance.soubory.delete()
+    if instance.historie and instance.historie.pk:
+        instance.historie.delete()
+    logger.debug("pas.signals.dokument_delete_container_soubor_vazby.end",
+                 extra={"ident_cely": instance.ident_cely})
 
 
 @receiver(pre_save, sender=UzivatelSpoluprace)
 def save_uzivatel_spoluprce(sender, instance: UzivatelSpoluprace, **kwargs):
     instance.vedouci.save_metadata(use_celery=False)
     instance.spolupracovnik.save_metadata(use_celery=False)
+
+
+@receiver(pre_delete, sender=UzivatelSpoluprace)
+def delete_uzivatel_spoluprce_connections(sender, instance: UzivatelSpoluprace, **kwargs):
+    if instance.historie and instance.historie.pk:
+        instance.historie.delete()
 
 
 @receiver(post_delete, sender=UzivatelSpoluprace)
